requirements.txt
# API Gmail
google-auth-oauthlib==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.0

# Interface utilisateur
PyQt6==6.6.1

# Traitement des données et IA
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
python-dotenv==1.0.0

# Utilitaires
python-dateutil==2.8.2
pydantic==2.5.2